Fs eth dev

+++ datahub/views.py
@@ -2,7 +2,7 @@ import json
 import logging
 import os
 import re
-import shutil
+import shutil, ast
 from calendar import c
 
 import django
@@ -1355,7 +1355,8 @@ class DatasetV2ViewSet(GenericViewSet):
                     Create below directories with dataset files uploaded
                     /temp/<dataset-name>/file/<files>
                 """
-                serializer = DatasetV2TempFileSerializer(data=request.data, context={"request_method": request.method})
+                # serializer = DatasetV2TempFileSerializer(data=request.data, context={"request_method": request.method})
+                serializer = DatasetV2TempFileSerializer(data=request.data, context={"request_method": request.method, "dataset_exists": request.query_params.get("dataset_exists"), "queryset": self.queryset})
                 if not serializer.is_valid():
                     return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
@@ -1410,10 +1411,11 @@ class DatasetV2ViewSet(GenericViewSet):
                     """Delete a single file as requested"""
                     file_name = request.data.get("file_name")
                     file_path = os.path.join(directory, request.data.get("source"), file_name)
-                    os.remove(file_path)
-                    LOGGER.info(f"Deleting file: {file_name}")
-                    data = {file_name: "File deleted"}
-                    return Response(data, status=status.HTTP_204_NO_CONTENT)
+                    if os.path.exists(file_path):
+                        os.remove(file_path)
+                        LOGGER.info(f"Deleting file: {file_name}")
+                        data = {file_name: "File deleted"}
+                        return Response(data, status=status.HTTP_204_NO_CONTENT)
 
                 return Response(status=status.HTTP_204_NO_CONTENT)
 
@@ -1464,6 +1466,43 @@ class DatasetV2ViewSet(GenericViewSet):
         serializer.save()
         return Response(serializer.data, status=status.HTTP_201_CREATED)
 
+    def update(self, request, pk, *args, **kwargs):
+        """
+        ``PUT`` method: PUT method to edit or update the dataset (DatasetV2) and its files (DatasetV2File). [see here][ref]
+
+        **Endpoint**
+        [ref]: /datahub/dataset/v2/<uuid>
+        """
+        setattr(request.data, "_mutable", True)
+        data = request.data
+        to_delete = ast.literal_eval(data.get("deleted", "[]"))
+        self.dataset_files(data, to_delete)
+        datasetv2 = self.get_object()
+        serializer = self.get_serializer(datasetv2, data=data, partial=True)
+        serializer.is_valid(raise_exception=True)
+        serializer.save()
+        return Response(serializer.data, status=status.HTTP_201_CREATED)
+
+    @action(detail=False, methods=["delete"])
+    def dataset_files(self, request, id=[]):
+        """
+        ``DELETE`` method: DELETE method to delete the dataset files (DatasetV2File) referenced by DatasetV2 model. [see here][ref]
+
+        **Endpoint**
+        [ref]: /datahub/dataset/v2/dataset_files/
+        """
+        ids = {}
+        for file_id in id:
+            dataset_file = DatasetV2File.objects.filter(id=int(file_id))
+            if dataset_file.exists():
+                LOGGER.info(f"Deleting file: {dataset_file[0].id}")
+                file_path = os.path.join("media", str(dataset_file[0].file))
+                if os.path.exists(file_path):
+                    os.remove(file_path)
+                dataset_file.delete()
+                ids[file_id] = "File deleted"
+        return Response(ids, status=status.HTTP_204_NO_CONTENT)
+
     def list(self, request, *args, **kwargs):
         """
         ``GET`` method Endpoint: list action to view the list of Datasets via GET request. [see here][ref].
@@ -1496,8 +1535,10 @@ class DatasetV2ViewSet(GenericViewSet):
         for file in dataset_file_obj:
             path_ = os.path.join("/media/", str(file.file))
             file_path = {}
+            file_path["id"] = file.id
             file_path["content"] = read_contents_from_csv_or_xlsx_file(path_)
             file_path["file"] = path_
+            file_path["source"] = file.source
             data.append(file_path)
 
         serializer["datasets"] = data
@@ -1578,6 +1619,29 @@ class DatasetV2ViewSet(GenericViewSet):
             return Response(f"Invalid filter fields: {list(request.data.keys())}", status=500)
         return Response({"geography": geography}, status=200)
 
+    def destroy(self, request, pk, *args, **kwargs):
+        """
+        ``DELETE`` method: DELETE method to delete the DatasetV2 instance and its reference DatasetV2File instances, 
+        along with dataset files stored at the URL. [see here][ref]
+
+        **Endpoint**
+        [ref]: /datahub/dataset/v2/
+        """
+        dataset_obj = self.get_object()
+        if dataset_obj:
+            dataset_files = DatasetV2File.objects.filter(dataset_id=dataset_obj.id)
+            dataset_dir = os.path.join(settings.DATASET_FILES_URL, str(dataset_obj.name))
+
+            if os.path.exists(dataset_dir):
+                shutil.rmtree(dataset_dir)
+                LOGGER.info(f"Deleting file: {dataset_dir}")
+
+            # delete DatasetV2File & DatasetV2 instances
+            LOGGER.info(f"Deleting dataset obj: {dataset_obj}")
+            dataset_files.delete()
+            dataset_obj.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+          
     @action(detail=False, methods=["post"])
     def search_datasets(self, request, *args, **kwargs):
         data = request.data
@@ -1621,6 +1685,4 @@ class DatasetV2ViewSet(GenericViewSet):
                 f"Invalid filter fields: {list(request.data.keys())}",
                 status=status.HTTP_500_INTERNAL_SERVER_ERROR,
             )
-
-    
-
+
